Rename pwndbg.aglib.disasm.arch to pwndbg.aglib.disasm.assistant
