+++ pwndbg/aglib/disasm/assistant.py
